+++ src/Microsoft.Maui.Graphics/IImage.cs
@@ -13,12 +13,12 @@ namespace Microsoft.Maui.Graphics
 
     public interface IImage : IDrawable, IDisposable
     {
-        float Width { get; }
-        float Height { get; }
-        IImage Downsize(float maxWidthOrHeight, bool disposeOriginal = false);
-        IImage Downsize(float maxWidth, float maxHeight, bool disposeOriginal = false);
-        IImage Resize(float width, float height, ResizeMode resizeMode = ResizeMode.Fit, bool disposeOriginal = false);
-        void Save(Stream stream, ImageFormat format = ImageFormat.Png, float quality = 1);
-        Task SaveAsync(Stream stream, ImageFormat format = ImageFormat.Png, float quality = 1);
+        double Width { get; }
+        double Height { get; }
+        IImage Downsize(double maxWidthOrHeight, bool disposeOriginal = false);
+        IImage Downsize(double maxWidth, double maxHeight, bool disposeOriginal = false);
+        IImage Resize(double width, double height, ResizeMode resizeMode = ResizeMode.Fit, bool disposeOriginal = false);
+        void Save(Stream stream, ImageFormat format = ImageFormat.Png, double quality = 1);
+        Task SaveAsync(Stream stream, ImageFormat format = ImageFormat.Png, double quality = 1);
     }
 }
